Test modules for making an observer into a tiny web service.
Based on a recipe in the Python Cookbook, 3rd Ed., this initial test
simply registers a function that reads the observer queue whenever
an HTTP GET request is received for that observer.

It's not refined, but it proves that we can access observer data remotely
via a REST interface.  This removes all the tedious "manager" crap I was
trying to dream up.

Hopefully, one process should be able to run all the desired metrics threads
and deliver their data as requested.

creates a single observation thread, then starts